feat(admin/users): refactor user management API and types into dedicated modules
 - Refactor user management API functions from `lib/actions.ts` to `lib/action/adminUserManagementApi.ts` for better modularity.
 - Extract Zod schemas and TypeScript types into a dedicated `lib/definition/adminUserManagementSchema.ts` file.
 - Create a new test suite (`adminUserManagementApi.test.ts`) with comprehensive coverage for the extracted API module.
 - Isolate MSW mock handlers for admin user endpoints into `mocks/admin/handlers.ts` to improve test organization.
 - Update import paths across all affected components and pages to reflect the new file structure.
 - Remove redundant tests from `lib/actions.test.ts` that are now covered in the new test file.
