+++ app/admin/users/[id]/page.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
-import { fetchUserDetails } from "@/lib/actions";
-import { AdminUserDetailResponse } from "@/lib/definition";
+import { fetchUserDetails } from "@/lib/action/adminUserManagementApi";
+import { AdminUserDetailResponse } from "@/lib/definition/adminUserManagementSchema";
 import EditButton from "@/components/admin/users/buttons/EditButton";
 import DeleteButton from "@/components/admin/users/buttons/DeleteButton";
