+++ components/admin/users/edit/AdminUserEditForm.tsx
@@ -2,7 +2,7 @@ import { LabelInput, ErrorText } from "@/components/common";
 import {
   AdminUserDetailResponse,
   AdminUserUpdateState,
-} from "@/lib/definition";
+} from "@/lib/definition/adminUserManagementSchema";
 import RoleSelect from "../RoleSelect";
 
 interface PageProps {
