+++ lib/actions.ts
@@ -9,7 +9,6 @@ import {
   LoginInfoState,
   NicknameUpdateSchema,
   NicknameUpdateState,
-  PaginationDto,
   PasswordUpdateSchema,
   PasswordUpdateState,
   PostResponse,
@@ -18,17 +17,6 @@ import {
   CreatePostInputState,
   UserSchema,
   UserState,
-  AdminUserListPaginatedSchema,
-  AdminUserListPaginatedResponse,
-  AdminUserCreateSchema,
-  AdminUserCreateState,
-  AdminUserCreate,
-  AdminUserUpdateSchema,
-  AdminUserUpdateState,
-  UserRole,
-  AdminUserDetailResponseSchema,
-  AdminUserDetailResponse,
-  SearchUserDtoSchema,
   CategoryWithBoardsResponse,
   CategoryWithBoardsResponseSchema,
 } from "./definition";
@@ -619,306 +607,4 @@ export async function FetchDashboardSummary(): Promise<DashboardSummaryDto> {
   }
 
   return result.data;
-}
-
-export async function FetchUserPaginatedList(
-  query: PaginationDto
-): Promise<AdminUserListPaginatedResponse> {
-  const { page = 1, limit = 10, sort = "createdAt", order = "DESC" } = query;
-  const params = new URLSearchParams({
-    page: page.toString(),
-    limit: limit.toString(),
-    sort,
-    order,
-  });
-
-  const reqAddress = `${serverAddress}/admin/users`;
-  const { data, error } = await authenticatedFetch({
-    method: "GET",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    url: `${reqAddress}?${params.toString()}`,
-  });
-
-  if (error) {
-    switch (error.type) {
-      case AuthenticatedFetchErrorType.Unauthorized:
-        await clearAuth();
-        redirect("/login?error=session_expired");
-      default:
-        console.error(
-          "사용자 목록을 불러오던 중 예기치 못한 오류 발생:",
-          error
-        );
-        throw new Error(
-          "사용자 목록을 불러오지 못했습니다. 잠시 후 다시 시도해주세요."
-        );
-    }
-  } else {
-    const result = AdminUserListPaginatedSchema.safeParse(data);
-    if (!result.success) {
-      console.error("Validation failed:", result.error);
-      throw new Error("Invalid data format for UserList");
-    }
-    return result.data;
-  }
-}
-
-export async function FetchUserSearchResults(query: {
-  searchEmail?: string;
-  searchNickname?: string;
-  searchRole?: UserRole;
-  page?: number;
-  limit?: number;
-  sort?: "createdAt" | "updatedAt";
-  order?: "ASC" | "DESC";
-}): Promise<AdminUserListPaginatedResponse> {
-  // Validate input using the schema
-  const validatedQuery = SearchUserDtoSchema.safeParse(query);
-
-  if (!validatedQuery.success) {
-    console.error("Invalid search parameters:", validatedQuery.error);
-    throw new Error("Invalid search parameters");
-  }
-
-  const {
-    page = 1,
-    limit = 10,
-    sort = "createdAt",
-    order = "DESC",
-    searchEmail,
-    searchNickname,
-    searchRole,
-  } = validatedQuery.data;
-
-  const params = new URLSearchParams({
-    page: page.toString(),
-    limit: limit.toString(),
-    sort,
-    order,
-  });
-
-  if (searchEmail) params.set("searchEmail", searchEmail);
-  if (searchNickname) params.set("searchNickname", searchNickname);
-  if (searchRole) params.set("searchRole", searchRole);
-
-  const reqAddress = `${serverAddress}/admin/users/search`;
-  const { data, error } = await authenticatedFetch({
-    method: "GET",
-    headers: { "Content-Type": "application/json" },
-    url: `${reqAddress}?${params.toString()}`,
-  });
-
-  if (error) {
-    switch (error.type) {
-      case AuthenticatedFetchErrorType.Unauthorized:
-        await clearAuth();
-        redirect("/login?error=session_expired");
-      default:
-        console.error("Failed to fetch search results:", error);
-        throw new Error("Failed to fetch search results");
-    }
-  } else {
-    const result = AdminUserListPaginatedSchema.safeParse(data);
-    if (!result.success) {
-      console.error("Validation failed:", result.error);
-      throw new Error("Invalid data format for search results");
-    }
-    return result.data;
-  }
-}
-
-export async function AdminSignUpAction(
-  prevState: AdminUserCreateState,
-  formData: FormData
-): Promise<AdminUserCreateState> {
-  const validatedFields = AdminUserCreateSchema.safeParse({
-    name: formData.get("name"),
-    nickname: formData.get("nickname"),
-    email: formData.get("email"),
-    password: formData.get("password"),
-    role: formData.get("role"),
-  });
-
-  if (!validatedFields.success) {
-    return {
-      errors: validatedFields.error.flatten().fieldErrors,
-      message: "Missing or invalid fields. Failed to create admin user.",
-    };
-  }
-
-  const userData = validatedFields.data;
-
-  const reqAddress = `${serverAddress}/admin/users`;
-
-  const { error } = await authenticatedFetch({
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(userData),
-    url: reqAddress,
-  });
-  if (error) {
-    switch (error.type) {
-      case AuthenticatedFetchErrorType.Unauthorized:
-        await clearAuth();
-        redirect("/login?error=session_expired");
-      case AuthenticatedFetchErrorType.ConflictError:
-        const message = error.message;
-        const fieldErrors: Partial<Record<keyof AdminUserCreate, string[]>> =
-          {};
-        if (message.includes("email")) {
-          fieldErrors.email = [ERROR_MESSAGES.EMAIL_EXISTS];
-        }
-        if (message.includes("nickname")) {
-          fieldErrors.nickname = [ERROR_MESSAGES.NICKNAME_EXISTS];
-        }
-
-        if (Object.keys(fieldErrors).length > 0) {
-          return {
-            message: ERROR_MESSAGES.INVALID_FIELD,
-            errors: fieldErrors,
-          };
-        }
-        return {
-          message: "Conflict occurred. Please check your input values.",
-        };
-      default:
-        console.error("Unexpected error during admin signup:", error);
-        return {
-          message: "An unexpected error occurred. Please try again.",
-        };
-    }
-  } else {
-    revalidatePath("/admin/users");
-    redirect("/admin/users");
-  }
-}
-
-export async function AdminUserUpdateAction(
-  userId: number,
-  prevState: AdminUserUpdateState,
-  formData: FormData
-): Promise<AdminUserUpdateState> {
-  const updateData = {
-    name: formData.get("name")?.toString().trim() || undefined,
-    nickname: formData.get("nickname")?.toString().trim() || undefined,
-    role: formData.get("role")?.toString().trim() as UserRole | undefined,
-  };
-
-  const validatedFields = AdminUserUpdateSchema.safeParse(updateData);
-
-  if (!validatedFields.success) {
-    return {
-      errors: validatedFields.error.flatten().fieldErrors,
-      message: "Invalid fields. Please check your input values.",
-    };
-  }
-
-  const reqAddress = `${serverAddress}/admin/users/${userId}`;
-
-  const { error } = await authenticatedFetch({
-    url: reqAddress,
-    method: "PATCH",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(validatedFields.data),
-  });
-
-  if (error) {
-    switch (error.type) {
-      case AuthenticatedFetchErrorType.Unauthorized:
-        await clearAuth();
-        redirect("/login?error=session_expired");
-      case AuthenticatedFetchErrorType.ConflictError:
-        const message = error.message;
-        const fieldErrors: Partial<Record<keyof AdminUserCreate, string[]>> =
-          {};
-        if (message.includes("nickname")) {
-          fieldErrors.nickname = [ERROR_MESSAGES.NICKNAME_EXISTS];
-        }
-
-        if (Object.keys(fieldErrors).length > 0) {
-          return {
-            message: ERROR_MESSAGES.INVALID_FIELD,
-            errors: fieldErrors,
-          };
-        }
-        return {
-          message: "Conflict occurred. Please check your input values.",
-        };
-      default:
-        console.error("Unexpected error during user manager editing:", error);
-        return {
-          message: "An unexpected error occurred. Please try again.",
-        };
-    }
-  } else {
-    revalidatePath("/admin/users");
-    redirect("/admin/users");
-  }
-}
-
-export async function fetchUserDetails(
-  id: number
-): Promise<AdminUserDetailResponse> {
-  const reqAddress = `${serverAddress}/admin/users/${id}`;
-
-  const { data, error } = await authenticatedFetch({
-    method: "GET",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    url: reqAddress,
-  });
-
-  if (error) {
-    switch (error.type) {
-      case AuthenticatedFetchErrorType.Unauthorized:
-        await clearAuth();
-        redirect("/login?error=session_expired");
-      case AuthenticatedFetchErrorType.NotFound:
-        throw new Error("사용자를 찾을 수 없습니다");
-      default:
-        throw new Error("사용자 정보를 불러오지 못했습니다");
-    }
-  }
-
-  // 응답 데이터 검증
-  const result = AdminUserDetailResponseSchema.safeParse(data);
-  if (!result.success) {
-    console.error("Validation failed:", result.error);
-    throw new Error("Invalid user data format");
-  }
-
-  return result.data;
-}
-
-export async function deleteUser(userId: number) {
-  const reqAddress = `${serverAddress}/admin/users/${userId}`;
-
-  const { error } = await authenticatedFetch({
-    method: "DELETE",
-    url: reqAddress,
-  });
-
-  if (error) {
-    console.error("User deletion failed:", error.message);
-    switch (error.type) {
-      case AuthenticatedFetchErrorType.Unauthorized:
-        await clearAuth();
-        redirect("/login?error=session_expired");
-      case AuthenticatedFetchErrorType.ServerError:
-        throw new Error("서버 오류가 발생했습니다.");
-      case AuthenticatedFetchErrorType.NotFound:
-        throw new Error("요청한 리소스를 찾을 수 없습니다.");
-      default:
-        throw new Error("사용자를 삭제할 수 없습니다.");
-    }
-  } else {
-    revalidatePath("/admin/users");
-  }
 }
